add remove from bloc link back fix legality switching
git-svn-id: https://svn.radicaldesigns.org/voterguides/trunk@233 75c0e4a9-b6e4-0310-a6ec-d70c6baf2916

+++ app/controllers/guides_controller.rb
@@ -343,8 +343,13 @@ class GuidesController < ApplicationController
 
   def unjoin
     pledge = Pledge.find_by_guide_id_and_user_id(params[:id], current_user.id)
-    pledge.destroy
-    render :partial => 'pledge', :locals => { :guide => @guide }, :layout => false
+    pledge.destroy if pledge
+    if request.xhr?
+      render :partial => 'pledge', :locals => { :guide => @guide }, :layout => false
+    else
+      @guide = pledge.guide
+      redirect_to guide_permalink_url(:year => @guide.date.year, :permalink => @guide.permalink)
+    end
   end
 
   def tell
